+++ tests/FactoryTest.php
@@ -4,10 +4,24 @@ use Clue\Redis\Protocol\Factory;
 
 class FactoryTest extends TestCase
 {
-    public function testCreate()
-    {
-        $protocol = Factory::create();
+    private $factory;
 
-        $this->assertInstanceOf('Clue\Redis\Protocol\ProtocolInterface', $protocol);
+    public function setUp()
+    {
+        $this->factory = new Factory();
+    }
+
+    public function testCreateParser()
+    {
+        $parser = $this->factory->createParser();
+
+        $this->assertInstanceOf('Clue\Redis\Protocol\Parser\ParserInterface', $parser);
+    }
+
+    public function testCreateSerializer()
+    {
+        $serializer = $this->factory->createSerializer();
+
+        $this->assertInstanceOf('Clue\Redis\Protocol\Serializer\SerializerInterface', $serializer);
     }
 }