+++ tests/Parser/RecursiveParserTest.php
@@ -1,0 +1,11 @@
+<?php
+
+use Clue\Redis\Protocol\Parser\RecursiveParser;
+
+class RecursiveParserTest extends AbstractParserTest
+{
+    protected function createProtocol()
+    {
+        return new RecursiveParser();
+    }
+}