+++ tests/Serializer/RecursiveSerializerTest.php
@@ -1,0 +1,11 @@
+<?php
+
+use Clue\Redis\Protocol\Serializer\RecursiveSerializer;
+
+class RecursiveSerializerTest extends AbstractSerializerTest
+{
+    protected function createSerializer()
+    {
+        return new RecursiveSerializer();
+    }
+}